third_party/blink/web_tests/storage/websql/sqlite-dbpage-not-allowed.html

<!doctype html>
<meta charset="utf-8">
<title>WebSQL: Use of the SQLITE_DBPAGE is not allowed</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

async_test(testCase => {
  const database = openDatabase(
      'SQLiteDbPageVTNotAllowedTest', '1.0', 'Database for SQLITE_DBPAGE blocking test',
      1024 * 1024);
  assert_not_equals(database, null, 'openDatabase should not fail');

  database.transaction(testCase.step_func(transaction => {
    transaction.executeSql(
        'DROP TABLE IF EXISTS main;', [], () => {},
        testCase.unreached_func('Table drop should not fail'));
    transaction.executeSql(
        'DELETE FROM sqlite_dbpage WHERE pgno=1;', [],
        testCase.unreached_func('sqlite_dbpage access should not succeed'),
        testCase.step_func_done());
    transaction.executeSql(
        'DELETE FROM SQLITE_DBPAGE WHERE pgno=1;', [],
        testCase.unreached_func('SQLITE_DBPAGE access should not succeed'),
        testCase.step_func_done());
    transaction.executeSql(
        'DELETE FROM sQlIte_DbPage WHERE pgno=1;', [],
        testCase.unreached_func('sQlIte_DbPage access should not succeed'),
        testCase.step_func_done());
    // Read access is allowed.
    transaction.executeSql('SELECT data FROM sqlite_dbpage WHERE pgno=1;');
  }));
}, 'writing to the SQLITE_DBPAGE virtual table should not be allowed');
</script>
